WIP: wire: Represent structs as a FieldList
Rather than holding a slice of fields, represent structs as a lazy
FieldList.

This behaves very similarly to the ValueList and MapItemList interfaces we
have.

This is strictly an internal change from the point of view of wire.Value.

TODO: Benchmark to see if this affects performance in any way.